MAINT: move to dtoolkit\geoaccessor\dataframe.py
