+++ dtoolkit/geoaccessor/__init__.py
@@ -1,3 +1,4 @@
+from dtoolkit.geoaccessor import dataframe  # noqa
 from dtoolkit.geoaccessor import geodataframe  # noqa
 from dtoolkit.geoaccessor import geoseries  # noqa
 from dtoolkit.geoaccessor.accessor import register_geodataframe_accessor  # noqa
